Fix broken corner case related to span collapsing
When deleting, re-inserting, and then deleting the same content,
the first two events were collapsed, but no span was added for
the third, because it technicall doesn't delete any original
content (the content it deleted had been inseted by the 2nd step).
This led to inconsistent state.

This patch drops the map history from the ChangeSet object, and
instead derives a map from the current spans when adding more
steps. That way, the mapping and the displayed spans are
consistent.

FIX: Fix issue where deleting, inserting, and deleting the same
content would lead to an inconsistent change set.

Issue newsdev/prosemirror-revision-browsing#30