Prevent greedy change merging across block nodes
With a large doc, the minUnchanged maxes to 15, meaning if
we have a change spanning 2 paragraphs the diistance between which
is less than 15, the diff will count this as a single replacement of
both paragraph. Instead, check if the spanned gap between the changes
contains a closing token (-1), if so do not combine the changes.